Serverless/requirements.txt
flask
torch
torchvision
google-cloud-storage
